Trait rome_analyze::GroupCategory

pub trait GroupCategory {
    type Language: Language;

    const CATEGORY: RuleCategory;

    fn record_groups<V: RegistryVisitor<Self::Language> + ?Sized>(
        registry: &mut V
    );
}

A group category is a collection of rule groups under a given category ID, serving as a broad classification on the kind of diagnostic or code action these rule emit, and allowing whole categories of rules to be disabled at once depending on the kind of analysis being performed

Required Associated Types

type Language: Language

Required Associated Constants

const CATEGORY: RuleCategory

The category ID used for all groups and rule belonging to this category

Required Methods

fn record_groups<V: RegistryVisitor<Self::Language> + ?Sized>(registry: &mut V)

Register all the groups belonging to this category into registry