Trait rome_analyze::GroupLanguage

pub trait GroupLanguage {
    type Language: Language;
}

This trait is implemented for tuples of Rule types of size 1 to 20 if the query type of all the rules in the tuple share the same associated [Language] (which is then aliased as the Language associated type on GroupLanguage itself). It is used to ensure all the rules in a given group are all querying the same underlying language

Required Associated Types

type Language: Language