Trait rome_analyze::NodeVisitor

pub trait NodeVisitor<V>: Sized {
    type Node: AstNode;

    fn enter(
        node: Self::Node,
        ctx: &mut VisitorContext<'_, '_, <Self::Node as AstNode>::Language>,
        stack: &mut V
    ) -> Self;
    fn exit(
        self,
        node: Self::Node,
        ctx: &mut VisitorContext<'_, '_, <Self::Node as AstNode>::Language>,
        stack: &mut V
    );
}

A node visitor is a special kind of visitor that does not have a persistent state for the entire run of the analyzer. Instead these visitors are transient, they get instantiated when the traversal enters the corresponding node type and destroyed when the corresponding node exits

Due to these specificities node visitors do not implement Visitor directly, instead one or more of these must the merged into a single visitor type using the crate::merge_node_visitors macro

Required Associated Types

type Node: AstNode

Required Methods

fn enter(
node: Self::Node,
ctx: &mut VisitorContext<'_, '_, <Self::Node as AstNode>::Language>,
stack: &mut V
) -> Self

fn exit(
self,
node: Self::Node,
ctx: &mut VisitorContext<'_, '_, <Self::Node as AstNode>::Language>,
stack: &mut V
)