Trait rome_analyze::Queryable

pub trait Queryable: Sized {
    type Output;
    type Language: Language;
    type Services: TryFrom<ServiceBag> + Phase;

    const KEY: QueryKey<Self::Language>;

    fn unwrap_match(query: &QueryMatch<Self::Language>) -> Self::Output;
}

Trait implemented for all types, for example lint rules can query them to emit diagnostics or code actions.

Required Associated Types

type Output

type Language: Language

type Services: TryFrom<ServiceBag> + Phase

Required Associated Constants

const KEY: QueryKey<Self::Language>

Statically declares which QueryMatch variant is matched by this Queryable type. For instance the Ast queryable matches on QueryMatch::Syntax, so its key is defined as QueryKey::Syntax

Required Methods

fn unwrap_match(query: &QueryMatch<Self::Language>) -> Self::Output

Unwrap an instance of Self from a QueryMatch.

Panics

If the QueryMatch variant of query doesn’t match Self::KEY