Trait rome_analyze::Rule

pub trait Rule: RuleMeta {
    type Query: Queryable;
    type State;
    type Signals: IntoIterator<Item = Self::State>;
    type Options: DeserializeOwned;

    fn run(ctx: &RuleContext<'_, Self>) -> Self::Signals;

    fn phase() -> Phases { ... }
    fn text_range(
        ctx: &RuleContext<'_, Self>,
        state: &Self::State
    ) -> Option<TextRange> { ... }
    fn suppressed_nodes(
        ctx: &RuleContext<'_, Self>,
        state: &Self::State,
        suppressions: &mut RuleSuppressions<<<Self as Rule>::Query as Queryable>::Language>
    ) { ... }
    fn diagnostic(
        _ctx: &RuleContext<'_, Self>,
        _state: &Self::State
    ) -> Option<RuleDiagnostic> { ... }
    fn action(
        ctx: &RuleContext<'_, Self>,
        state: &Self::State
    ) -> Option<RuleAction<<<Self as Rule>::Query as Queryable>::Language>> { ... }
}

Trait implemented by all analysis rules: declares interest to a certain AstNode type, and a callback function to be executed on all nodes matching the query to possibly raise an analysis event

Required Associated Types

type Query: Queryable

The type of AstNode this rule is interested in

type State

A generic type that will be kept in memory between a call to run and subsequent executions of diagnostic or action, allows the rule to hold some temporary state between the moment a signal is raised and when a diagnostic or action needs to be built

type Signals: IntoIterator<Item = Self::State>

An iterator type returned by run to yield zero or more signals to the analyzer

type Options: DeserializeOwned

The options that belong to a rule

Required Methods

fn run(ctx: &RuleContext<'_, Self>) -> Self::Signals

This function is called once for each node matching Query in the tree being analyzed. If it returns Some the state object will be wrapped in a generic AnalyzerSignal, and the consumer of the analyzer may call diagnostic or action on it

Provided Methods

fn phase() -> Phases

fn text_range(
ctx: &RuleContext<'_, Self>,
state: &Self::State
) -> Option<TextRange>

Used by the analyzer to associate a range of source text to a signal in order to support suppression comments.

If this function returns None, the range of the query node will be used instead

The default implementation returns the range of Self::diagnostic, and should return the correct value for most rules however you may want to override this if generating a diagnostic for this rule requires heavy processing and the range could be determined through a faster path

fn suppressed_nodes(
ctx: &RuleContext<'_, Self>,
state: &Self::State,
suppressions: &mut RuleSuppressions<<<Self as Rule>::Query as Queryable>::Language>
)

Allows the rule to suppress a set of syntax nodes to prevent them from matching the Query. This is useful for rules that implement a code action that recursively modifies multiple nodes at once, this hook allows these rules to avoid matching on those nodes again.

Example

impl Rule for SimplifyExpression {
    type Query = BinaryExpression;

    fn run(ctx: &RuleContext<Self>) -> Self::Signals {
        // Recursively check this expression and its children for simplification
        // opportunities
        check_can_simplify(ctx.query())
    }

    fn suppressed_nodes(
        _ctx: &RuleContext<Self>,
        state: &Self::State,
        suppressions: &mut RuleSuppressions<RuleLanguage<Self>>
    ) {
        // Prevent this rule from matching again on nodes that were already checked by
        // `check_can_simplify`
        for node in &state.nodes {
            suppressions.suppress_node(node.clone());
        }
    }
}

fn diagnostic(
_ctx: &RuleContext<'_, Self>,
_state: &Self::State
) -> Option<RuleDiagnostic>

Called by the consumer of the analyzer to try to generate a diagnostic from a signal raised by run

The default implementation returns None

fn action(
ctx: &RuleContext<'_, Self>,
state: &Self::State
) -> Option<RuleAction<<<Self as Rule>::Query as Queryable>::Language>>

Called by the consumer of the analyzer to try to generate a code action from a signal raised by run

The default implementation returns None