Trait rome_analyze::Visitor

pub trait Visitor {
    type Language: Language;

    fn visit(
        &mut self,
        event: &WalkEvent<SyntaxNode<Self::Language>>,
        ctx: VisitorContext<'_, '_, Self::Language>
    );

    fn finish(self: Box<Self>, ctx: VisitorFinishContext<'_, Self::Language>) { ... }
}

Visitors are the main building blocks of the analyzer: they receive syntax WalkEvents, process these events to build secondary data structures from the syntax tree, and emit rule query matches through the crate::RuleRegistry

Required Associated Types

type Language: Language

Required Methods

fn visit(
&mut self,
event: &WalkEvent<SyntaxNode<Self::Language>>,
ctx: VisitorContext<'_, '_, Self::Language>
)

Provided Methods

fn finish(self: Box<Self>, ctx: VisitorFinishContext<'_, Self::Language>)