Type Definition rome_analyze::ControlFlow

pub type ControlFlow<B = Never> = ControlFlow<B>;

Type alias of ops::ControlFlow with the B generic type defaulting to Never

By default the analysis loop never breaks, so it behaves mostly like let b = loop {}; and has a “break type” of ! (the ! type isn’t stable yet so I’m using an empty enum instead but they’re identical for this purpose)

In practice it’s not really a loop but a for because it’s iterating on all nodes in the syntax tree, so when it reaches the end of the iterator the loop will exit but without producing a value of type B: for this reason the analyze function returns an Option<B> that’s set to Some(B) if the callback did break, and None if the analysis reached the end of the file.

Most consumers of the analyzer will want to analyze the entire file at once and never break, so using Never as the type of B in this case lets the compiler know the ControlFlow::Break branch will never be taken and can be optimized out, as well as completely remove the return Some case (Option<Never> has a size of 0 and can be elided, while Option<()> has a size of 1 as it still need to store a discriminant)