Crate rome_analyze

Analyzer

The analyzer is a generic crate aimed to implement a visitor-like infrastructure, where it’s possible to inspect a piece of AST and emit diagnostics or actions based on a static check.

Folder structure

First of all, you need to identify the crate where you want to implement the rule. If the rule is going to be implemented for the JavaScript language (and its super languages), then the rule will be implemented inside the rome_js_analyzer crate.

Rules are divided by capabilities:

analyzers/ folder contains rules that don’t require any particular capabilities;

semantic_analyzer/ folder contains rules that require the use of the semantic model;

assists/ folder contains rules that contribute to refactor code, with not associated diagnostics; these are rules that are usually meant for editors/IDEs;

Most of the rules will go under semantic_analyzer/ or analyzers/ folders.

Inside the folders, we will have folders for each group that Rome supports.

When implementing new rules, they have to be implemented under the group nursery. New rules should always be considered unstable/not exhaustive.

Lint rules

This gives to the project time to test the rule, find edge cases, etc.

When creating or updating lint rules, you need to be aware that there’s a lot of generated code inside our toolchain, and our CI makes sure that this code is not out of sync. If some autogenerated files are out of sync, the CI will fail.

Create your first rule

Let’s say we want to create a new rule called useAwesomeTricks, which uses the semantic model.

create a new file under semantic_analyzers/nursery called use_awesome_tricks;

run the cargo alias cargo codegen analyzer, this command will update the file called nursery.rs inside the semantic_analyzers folder

from there, use the declare_rule macro to create a new type

use rome_analyze::declare_rule;
 
declare_rule! {
  /// Promotes the use of awesome tricks
  /// 
  /// ## Examples
  ///
  /// ### Invalid
  ///
  pub(crate) UseAwesomeTricks {
      version: "0.10.0",
      name: "useAwesomeTricks",
      recommended: false,
     }
 }

Then you need to use the Rule trait to implement the rule on this new created struct

use rome_analyze::{Rule, RuleCategory};
use rome_js_syntax::JsAnyExpression;
use rome_analyze::context::RuleContext;

impl Rule for UseAwesomeTricks {
     const CATEGORY: RuleCategory = RuleCategory::Lint;
     type Query = Semantic<JsAnyExpression>;
     type State = String;
     type Signals = Option<Self::State>;
     type Options = ();

     fn run(ctx: &RuleContext<Self>) -> Self::Signals {}
}

the const CATEGORY must be RuleCategory::Lint otherwise it won’t work

the Query needs to have the Semantic type, because we want to have access to the semantic model. Query tells the engine on which AST node we want to trigger the rule.

The State type doesn’t have to be used, so it can be considered optional, but it has be defined as type State = ()

The run function must be implemented. This function is called every time the analyzer finds a match for the query specified by the rule, and may return zero or more “signals”.

Implement the optional diagnostic function, to tell the user where’s the error and why:

impl Rule for UseAwesomeTricks {
    // .. code
    fn diagnostic(_ctx: &RuleContext<Self>, _state: &Self::State) -> Option<RuleDiagnostic> {}
}

While implementing the diagnostic, please keep Rome’s technical principals in mind. This function is called of every signal emitted by the run function, and it may return zero or one diagnostic.

You will have to manually update the file rome_diagnostics_categories/src/categories.rs and add a new category for the new rule you’re about to create.

Implement the optional action function, if we are able to provide automatic code fix to the rule:

impl Rule for UseAwesomeTricks {
    // .. code
    fn action(_ctx: &RuleContext<Self>, _state: &Self::State) -> Option<JsRuleAction> {}
}

This function is called of every signal emitted by the run function, and it may return zero or one code action. When returning a code action, you will need to pass category and applicability fields. category must be ActionCategory::QuickFix, while applicability must be Applicability:MaybeIncorrect

That’s it! Now, let’s test the rule

Test a rule

Inside the tests/spec folder, rules are divided by group and rule name. The test infrastructure is rigid around the association of the pair “group/rule name”, which means that your test cases are placed inside the wrong group, you won’t see any diagnostics.

Since each new rule will start from nursery, that’s where we start.

From there, you have two options:

create a single file called like the rule name, e.g. useAwesomeTricks.js, useAwesomeTricks.tsx, etc. The extension of the file matters based on which kind of file you need to test;

create a folder called useAwesomeTricks/, and then create various files where you want to create different cases. These options are useful if your rules target different super languages, or you want to split your cases among different files;

Run the command cargo test and if you’ve done everything correctly, you should see some snapshots emitted with diagnostics and code actions.

Check our main contribution document to know how to deal with the snapshot tests.

Code generation

Run the following commands to update the generated files:

cargo codegen-configuration, this command must be run first and, it will update the configuration;

cargo lintdoc, it will update the website with the documentation of the rules, check declare_rule for more information about it;

cargo codegen-bindings, it will update the TypeScript types released inside the JS APIs;

cargo codegen-schema, it will update the JSON Schema file of the configuration, used by VSCode;

Naming patterns for rules

Forbid a concept

no<Concept>

When a rule’s sole intention is to forbid a single concept - such as disallowing the use of debugger statements - the rule should be named using the no prefix. For example, the rule to disallow the use of debugger statements is named noDebugger.

Mandate a concept

use<Concept>

When a rule’s sole intention is to mandate a single concept - such as forcing the use of camel-casing - the rule should be named using the use prefix. For example, the rule to mandating the use of camel-cased variable names is named useCamelCase.

declare_rule

This macro is used to declare an analyzer rule type, and implement the RuleMeta trait for it

Example

The macro itself expect the following syntax:

use rome_analyze::declare_rule;

declare_rule! {
    /// Documentation
    pub(crate) ExampleRule {
        version: "0.7.0",
        name: "ruleName",
        recommended: false,
    }
}

Documentation

The doc-comment for the rule is mandatory and is used to generate the documentation page for the rule on the website.

Importantly, the tool used to generate those pages also runs tests on the code blocks included in the documentation written in languages supported by the Rome toolchain (JavaScript, JSX, TypeScript, …) similar to how rustdoc generates tests from code blocks written in Rust. Because code blocks in Rust doc-comments are assumed to be written in Rust by default the language of the test must be explicitly specified, for instance:

use rome_analyze::declare_rule;
declare_rule! {
    /// Disallow the use of `var`
    ///
    /// ### Valid
    ///
    /// ```js
    /// let a, b;
    /// ```
    pub(crate) NoVar {
        version: "0.7.0",
        name: "noVar",
        recommended: false,
    }
}

Additionally, it’s possible to declare that a test should emit a diagnostic by adding expect_diagnostic to the language metadata:

use rome_analyze::declare_rule;
declare_rule! {
    ///  Disallow the use of `var`
    /// 
    ///  ### Invalid
    /// 
    ///  ```js,expect_diagnostic
    ///  var a, b;
    ///  ```
    pub(crate) NoVar {
        version: "0.7.0",
        name: "noVar",
        recommended: false,
    }
}

This will cause the documentation generator to ensure the rule does emit exactly one diagnostic for this code, and to include a snapshot for the diagnostic in the resulting documentation page

Deprecation

There are occasions when a rule must be deprecated, to avoid breaking changes. The reason of deprecations can be multiples.

In order to do, the macro allows to add additional field to add the reason for deprecation

use rome_analyze::declare_rule;

declare_rule! {
     /// Disallow the use of `var`
     /// 
     /// ### Invalid
     /// 
     /// ```js,expect_diagnostic
     /// var a, b;
     /// ```
    pub(crate) NoVar {
        version: "0.7.0",
        name: "noVar",
        deprecated: "Use the rule `noAnotherVar`",
        recommended: false,
    }
}

Category Macro

Declaring a rule using declare_rule! will cause a new rule_category! macro to be declared in the surrounding module. This macro can be used to refer to the corresponding diagnostic category for this lint rule, if it has one. Using this macro instead of getting the category for a diagnostic by dynamically parsing its string name has the advantage of statically injecting the category at compile time and checking that it is correctly registered to the rome_diagnostics library

declare_rule! {
    /// Documentation
    pub(crate) ExampleRule {
        version: "0.7.0",
        name: "ruleName"
    }
}

impl Rule for ExampleRule {
    fn diagnostic(ctx: &RuleContext<Self>, _state: &Self::State) -> Option<RuleDiagnostic> {
        Some(RuleDiagnostic::new(
            rule_category!(),
            ctx.query().text_trimmed_range(),
            "message",
        ))
    }
}

Modules

context

Macros

category_concat

The category_concat! macro is a variant of category! using a slightly different syntax, for use in the declare_group and declare_rule macros in the analyzer

declare_category

declare_group

This macro is used by the codegen script to declare an analyzer rule group, and implement the RuleGroup trait for it

declare_rule

This macro is used to declare an analyzer rule type, and implement the

merge_node_visitors

Creates a single struct implementing Visitor over a collection of type implementing the NodeVisitor helper trait. Unlike the global Visitor, node visitors are transient: they get instantiated when the traversal enters the corresponding node and destroyed when the node is exited. They are intended as a building blocks for creating and managing the state of complex visitors by allowing the implementation to be split over multiple smaller components.

Structs

AnalysisFilter

Allows filtering the list of rules that will be executed in a run of the analyzer, and at what source code range signals (diagnostics or actions) may be raised

Analyzer

The analyzer is the main entry point into the rome_analyze infrastructure. Its role is to run a collection of Visitors over a syntax tree, with each visitor implementing various analysis over this syntax tree to generate auxiliary data structures as well as emit “query match” events to be processed by lint rules and in turn emit “analyzer signals” in the form of diagnostics, code actions or both

AnalyzerAction

Code Action object returned by the analyzer, generated from a crate::RuleAction with additional information about the rule injected by the analyzer

AnalyzerConfiguration

A data structured derived from the rome.json file

AnalyzerContext

AnalyzerOptions

A set of information useful to the analyzer infrastructure

AnalyzerRules

A convenient new type data structure to insert and get rules

Ast

Query type usable by lint rules to match on specific AstNode types

InspectMatcher

Adapter type wrapping a QueryMatcher type with a function that can be used to inspect the query matches emitted by the analyzer

MatchQueryParams

Parameters provided to QueryMatcher::match_query and require to run lint rules

MetadataRegistry

Stores metadata information for all the rules in the registry, sorted alphabetically

MissingServicesDiagnostic

RegistryRuleMetadata

Metadata entry for a rule and its group in the registry

RuleAction

Code Action object returned by a single analysis rule

RuleCategories

RuleDiagnostic

Diagnostic object returned by a single analysis rule

RuleKey

Opaque identifier for a single rule

RuleMetadata

Static metadata containing information about a rule

RuleRegistry

The rule registry holds type-erased instances of all active analysis rules for each phase. What defines a phase is the set of services that a phase offers. Currently we have:

RuleRegistryBuilder

RuleSuppressions

Set of nodes this rule has suppressed from matching its query

ServiceBag

SignalEntry

Entry for a pending signal in the signal_queue

SyntaxVisitor

The SyntaxVisitor is the simplest form of visitor implemented for the analyzer, it simply broadcast each WalkEvent::Enter as a query match event for the SyntaxNode being entered

VisitorContext

Mutable context objects shared by all visitors

VisitorFinishContext

Mutable context objects provided to the finish hook of visitors

Enums

ActionCategory

AnalyzerDiagnostic

Small wrapper for diagnostics during the analysis phase.

Never

Utility type to be used as a default value for the B generic type on analyze when the provided callback never breaks

Phases

Defines all the phases that the RuleRegistry supports.

QueryKey

Mirrors the variants of QueryMatch to statically compute which queries a given Queryable type can match

QueryMatch

Enumerate all the types of Queryable analyzer visitors may emit

RuleCategory

RuleFilter

Allow filtering a single rule or group of rules by their names

Traits

AnalyzerSignal

Event raised by the analyzer when a Rule emits a diagnostic, a code action, or both

CategoryLanguage

This trait is implemented for tuples of Rule types of size 1 to 29 if the language of all the groups in the tuple share the same associated Language (which is then aliased as the Language associated type on CategoryLanguage itself). It is used to ensure all the groups in a given category are all querying the same underlying language

FromServices

GroupCategory

A group category is a collection of rule groups under a given category ID, serving as a broad classification on the kind of diagnostic or code action these rule emit, and allowing whole categories of rules to be disabled at once depending on the kind of analysis being performed

GroupLanguage

This trait is implemented for tuples of Rule types of size 1 to 29 if the query type of all the rules in the tuple share the same associated Language (which is then aliased as the Language associated type on GroupLanguage itself). It is used to ensure all the rules in a given group are all querying the same underlying language

NodeVisitor

A node visitor is a special kind of visitor that does not have a persistent state for the entire run of the analyzer. Instead these visitors are transient, they get instantiated when the traversal enters the corresponding node type and destroyed when the corresponding node exits

Phase

Defines which phase a rule will run. This will be defined by the set of services a rule demands.

QueryMatcher

The QueryMatcher trait is responsible of running lint rules on QueryMatch instances emitted by the various Visitor and push signals wrapped in SignalEntry to the signal queue

Queryable

Trait implemented for all types, for example lint rules can query them to emit diagnostics or code actions.

RegistryVisitor

Rule

Trait implemented by all analysis rules: declares interest to a certain AstNode type, and a callback function to be executed on all nodes matching the query to possibly raise an analysis event

RuleGroup

A rule group is a collection of rules under a given name, serving as a “namespace” for lint rules and allowing the entire set of rules to be disabled at once

RuleMeta

Visitor

Visitors are the main building blocks of the analyzer: they receive syntax WalkEvents, process these events to build secondary data structures from the syntax tree, and emit rule query matches through the crate::RuleRegistry

Type Definitions

ControlFlow

Type alias of ops::ControlFlow with the B generic type defaulting to Never

LanguageRoot