Macro rome_analyze::merge_node_visitors

macro_rules! merge_node_visitors {
    ( $vis:vis $name:ident { $( $id:ident: $visitor:ty, )+ } ) => { ... };
}

Creates a single struct implementing Visitor over a collection of type implementing the NodeVisitor helper trait. Unlike the global Visitor, node visitors are transient: they get instantiated when the traversal enters the corresponding node and destroyed when the node is exited. They are intended as a building blocks for creating and managing the state of complex visitors by allowing the implementation to be split over multiple smaller components.

Example

struct BinaryVisitor;

impl NodeVisitor for BinaryVisitor {
    type Node = BinaryExpression;
}

struct UnaryVisitor;

impl NodeVisitor for UnaryVisitor {
    type Node = UnaryExpression;
}

merge_node_visitors! {
    // This declares a new `ExpressionVisitor` struct that implements
    // `Visitor` and manages instances of `BinaryVisitor` and
    // `UnaryVisitor`
    pub(crate) ExpressionVisitor {
        binary: BinaryVisitor,
        unary: UnaryVisitor,
    }
}